requirements.txt
Django==6.0
djangorestframework==3.16.1
openai==2.14.0
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
Pillow>=10.0
gunicorn>=21.2
whitenoise>=6.6
